hopefully got it?
I replaced parser with ReplacementSort which runs my replacement sort. I forgot to change the arraylist for runlengths on public or make a getter for it, so if you use this you will have to change that. ReplacementSort is just a heavily modified Parser that contains all the steps for generating the runs and such. I also added a heap class in a separate file that you might want to check over. I am almost 100% sure that should work perfectly.